server/requirements.txt
flask-restplus==0.12.1
flask-cors==3.0.8
Flask-Migrate==2.5.2
psycopg2==2.8.3
SQLAlchemy==1.3.6